Registration File Update
This is an update to the registration.html file.

I updated the style.css to make the form stand out a bit and highlight the current field and a few other little things.

I also went through and programmed formatting checkers in JS for fields that are called when the HTML field loses focus. This includes a regex that just peels the numbers out of the phone number field and replaces them with a preformatted phone number. I also changed the date field to a select menu to allow us to have much more form control and not have to worry about the complexities of date formatting. The other thing I added was a password checker to make sure that the two passwords entered are equal.

Currently it just puts up an alert box when something isn't right. We should definitely change this, but I think it's low priority in comparison to the rest of the project.